SYSS-1160 Refactor Adapters to use Purse (#4)
* SYSS-1160 Add purse

* SYSS-1160 Add helper functions for World

* SYSS-1160 Add helper functions for World

* SYSS-1160 Rename functions

+++ mondex/dsl/src/main/kotlin/com/anaplan/engineering/azuki/mondex/dsl/declaration/MondexDeclarations.kt
@@ -3,7 +3,7 @@ package com.anaplan.engineering.azuki.mondex.dsl.declaration
 import com.anaplan.engineering.azuki.mondex.dsl.WorldBlock
 
 interface PurseDeclarations {
-    fun thereIsAPurse(balance: ULong, lost: ULong)
+    fun thereIsAPurse(balance: Int, lost: Int)
 }
 
 interface WorldDeclarations {
